pyproject.toml
[project]
name = "debt-by-ai"
version = "0.1.0"
description = "A game where you have to convince AI to get into debt, or have to convince AI to put you out of debt."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "flask>=3.1.2",
    "google-genai>=1.40.0",
    "python-dotenv>=1.1.1",
    "requests>=2.32.5",
]
